refactor(pi-claude-bridge): split index.ts along module seams (3.1.0)
Pure moves — no behavior changes; the bundle's export surface is verified
unchanged (99 exports, before/after diff clean; single cc-session-io chunk).

E1 consumeQuery + ClaudeAttemptFailure/ConsumeQueryResult +
   noteFastModeDisabledReason + emitRateLimitEvent → src/consume-query.ts.
E2 probeClaudeAccountProfile + BRIDGE_ACCOUNT_HOST +
   ACCOUNT_PROBE_DEADLINE_MS → src/account-host.ts; the shared mutable
   sdkQueryFactory test seam → src/sdk-query.ts (setter still re-exported
   from the entry).
E3 Pure buildClaudeQueryOptions() (SDK query-option assembly, effort
   mapping, resolveConfiguredEffort) → src/query-options.ts.
E4 Command/settings UI block (commandCwd, tryOpenExtensionManagerSettings,
   showBridgeStatus, reportConnectorInventory, registerBridgeCommands) →
   src/bridge-commands.ts.

index.ts: 1921 → 1347 lines; the streamClaudeAgentSdk closure web itself is
deliberately untouched.

Packaging: version 3.0.1 → 3.1.0 with a consumer-impacting CHANGELOG entry
covering the C1/C2 session semantics, the C13 builtin allowlist, the C14
enableConnectors/connectorWriteMode user-scope migration, C6 prime retries,
C15 credential scoping, and the module split; README + DEVELOPMENT document
the allowlist and the user-scope-only connector keys.

Validated: typecheck clean, 453 unit tests green, full npm test green
including live integration (smoke, multi-turn, cache, cursor-after-tools,
fork, session-compact, session-new, session-rebuild, session-resume,
tool-message).
