+++ pi-extensions/pi-claude-bridge/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Consumer-impacting changes
 
+### 3.1.0
+
+- **Connectors mode now enforces a fail-closed builtin ALLOWLIST at call time.** Every connectors-mode child carries a `PreToolUse` hook that permits only Pi's bridged custom tools (`mcp__custom-tools__*`), claude.ai connector tools (`mcp__claude_ai_*`, writes still subject to the write-deny gate), and the discovery built-ins (`ToolSearch`, `ListMcpResources`, `ReadMcpResource`) — and denies everything else, including future CLI built-ins the static denylist has never heard of. The hook applies in BOTH write modes (a one-shot write executor is still a connectors session ingesting third-party content), and every exception path inside it converts to a deny. New exports: `connectorBuiltinAllowlistHook`, `isAllowlistedConnectorSessionTool`, `denyAllToolsHook`.
+- **MIGRATION: `enableConnectors` and `connectorWriteMode` now resolve from user-scope settings and env only.** A project's checked-in `.pi/settings.json` or `.pi/claude-bridge.json` can no longer flip them, even for a trusted project — connectors expose live account data, so a cloned repo must not be able to switch them on (or un-gate writes) by being the cwd. If you previously enabled connectors per-project, move the keys to `<PI_CODING_AGENT_DIR>/settings.json` (or `claude-bridge.json`), or set `CLAUDE_BRIDGE_ENABLE_CONNECTORS`/`CLAUDE_BRIDGE_CONNECTOR_WRITE` in the process env. All other options keep the existing trusted-project layering.
+- **Reentrant (subagent) queries no longer touch the shared Claude session.** A subagent call routed through the parent's provider instance used to be able to shrink the parent's cursor to its own short context, clamp into a REUSE plan that resumed the parent's Claude session, or overwrite the parent's session record on completion. The prompt-batch planner now rejects a cursor beyond the messages array as proof of foreign context, the delivery-path cursor can only advance, and a reentrant fresh query runs as a clean one-shot: no resume id from the parent, no shared-session writes.
+- **Deferred (mid-query) user messages are never silently lost on failure paths (#967 invariant).** Every path that discards captured steers now records a `deferred_user_messages_dropped` diagnostic (count + previews), and any session record that survives the drop is persisted with `needsRebuild` so the next turn re-imports the steers from Pi history instead of resuming past them forever. Previously a terminal failure persisted the advanced cursor with no rebuild, losing the steers permanently.
+- Terminal `result` failures and late connector results arriving after a tool-use turn boundary (Pi stream already ended) are now captured: failures classify for account routing instead of masquerading as success, and a connector result observed late is audited `ok`/`error` instead of `unobserved`.
+- A managed rate limit whose SDK iterator then throws is counted once: the router's `recordRateLimit` cooldown is no longer doubled by a re-classified `recordFailure` for the same rejection. Companion-router callbacks are also exception-isolated — a throwing telemetry callback can no longer fail a delivered turn.
+- A failed connector-inventory prime (network blip, missing credentials) is no longer cached for the process lifetime; the next turn retries, keeping declarations and the disk-cache fallback reachable.
+- A managed profile with a missing `.credentials.json` no longer silently borrows the DEFAULT account's OAuth token: when a credential dir is explicitly selected, only that dir is probed (`credentialCandidatePaths`).
+- The account probe child (`probeProfile`) now runs under full bridge tool isolation plus a deny-all `PreToolUse` hook — it previously spawned with `bypassPermissions` and no tool restrictions.
+- Non-success terminal subtypes (`error_max_turns`, `error_during_execution`) surface with session bookkeeping intact as before, and the session-verify warning toast now home-relativizes paths (the local diagnostic dump keeps absolutes).
+- New exports: `scopeKeyFor` (the canonical credential-scope key rule shared with the connector cache's on-disk format) and `connectorServersSnapshot`.
+- Internal module split: `index.ts` shed the SDK-stream consumer, account host/probe, query-option assembly, connector runtime, and command registration into dedicated modules. The bundle's export surface is unchanged (verified by diffing the export list before/after).
+
 ### 3.0.1
 
 - **Connectors mode now loads only USER-scope settings in the child Claude Code** (previously `["user", "project", "local"]`). Claude Code settings files can set an `env` map and `apiKeyHelper`, so a checkout's `.claude/settings.json` could reintroduce the provider-override env the bridge scrubs from the child (e.g. `ANTHROPIC_BASE_URL`) on any connector query run in that cwd. Connector state lives in user scope, so connectors still surface unchanged. If a session genuinely needs project/local settings in the child, set `provider.settingSources` explicitly in bridge config (it still wins verbatim) — knowing it reopens the repo-controlled settings surface. Non-connectors behavior is unchanged. New export: `settingSourcesForQuery`.
