+++ pi-extensions/pi-claude-bridge/DEVELOPMENT.md
@@ -102,6 +102,8 @@ they are the only consumers). What the original evaluation flagged, and how each
 
 `loadConfig` layers three sources, lowest precedence first: `<piUserDir>/claude-bridge.json`, a trusted project's `.pi/claude-bridge.json`, then extension-manager config in `settings.json`. Isolated mode (`CLAUDE_BRIDGE_ISOLATED=1`) keeps only the first.
 
+Two provider keys are exempt from project layering: `enableConnectors` and `connectorWriteMode` resolve from **user scope and the env vars only** (`USER_SCOPE_ONLY_PROVIDER_KEYS` in `src/config.ts`). Project trust covers ordinary options; a checked-in `.pi/settings.json` or `.pi/claude-bridge.json` must not be able to expose the account's live connectors or un-gate their writes by being the cwd — the same posture as the setting-source lockdown below.
+
 Each `claude-bridge.json` is read by `legacyFileConfig`, which accepts both the nested legacy shape (`provider.*`, `promptContext.*`) and the manager's flat manifest keys; nested wins when a file carries both for one key. Provider values are normalized once over the merged result, so an invalid `forceEffort` in a higher layer still clears a valid one below it.
 
 `resolveExternalConfigValue(key, cwd)` reports what those files — and only those files — resolve for one manifest key, plus the concrete file that supplied it. It shares `legacyLayers`/`mergeLayers` with `loadConfig` and applies the same normalization, so the two cannot drift. `registerExternalConfigResolver` publishes it under `Symbol.for("vstack.pi.extension-config-resolver")` keyed by `PACKAGE_ID`; the vstack extension manager calls it when neither of its own scopes holds a key, and renders the value with its source file. Registration happens before the `config.enabled === false` early return, because a bridge disabled by `claude-bridge.json` is precisely the case the settings editor has to explain. The contract itself is documented in [`pi-extension-manager/DEVELOPMENT.md`](../pi-extension-manager/DEVELOPMENT.md).
@@ -112,6 +114,8 @@ Write denial with `connectorWriteMode: "deny"` is two-layered:
 
 - **Model context:** the known write tools are passed as `disallowedTools` (exact tool ids). The CLI's MCP permission matcher only supports exact tool names or a whole-server `mcp__server__*` glob — partial tool-segment globs are inert — so exact ids are what actually removes today's writes.
 - **Runtime:** a `PreToolUse` hook blocks any connector tool classified as a write at call time, regardless of permission mode. Classification is fail-closed over the whole `mcp__claude_ai_<Server>__` space: a tool there is a write unless its name *begins* with a known read verb (`list`, `search`, `get`, `read`, `fetch`, …). The verb is matched as a word across naming styles — `search_threads` (Gmail), `slack_read_channel` (Slack, server-prefixed), and `getJiraIssue` (Atlassian, camelCase) are all reads; a leading word that merely repeats the server name is skipped first. A name that opens with a read verb but also names a mutation (`getOrCreateChannel`) is a write, and so is a name that does not parse as `<server>__<tool>`.
+
+Independent of write mode, every connectors-mode child also carries a fail-closed **builtin allowlist** `PreToolUse` hook (`connectorBuiltinAllowlistHook` in `src/connectors.ts`): only `mcp__custom-tools__*`, `mcp__claude_ai_*`, and the discovery built-ins (`CONNECTOR_DISCOVERY_TOOLS`) may execute; everything else — including CLI built-ins that postdate `DISALLOWED_BUILTIN_TOOLS` — is denied, and a hook exception converts to a deny (the CLI treats a thrown hook as empty output, i.e. fail-open, so every catch path must deny). The account probe child gets the harsher `denyAllToolsHook` — a `/usage` probe executes nothing.
 
 ## Connector setting-source lockdown
