+++ pi-extensions/pi-claude-bridge/src/index.ts
@@ -1,22 +1,19 @@
 import { type AssistantMessage, type AssistantMessageEventStream, type Context, type Model, type SimpleStreamOptions, type Tool } from "@earendil-works/pi-ai";
 import * as piAi from "@earendil-works/pi-ai";
-import { type ExtensionAPI, type ExtensionUIContext } from "@earendil-works/pi-coding-agent";
-import { createSdkMcpServer, query, type EffortLevel, type SDKUserMessage, type SettingSource } from "@anthropic-ai/claude-agent-sdk";
+import { type ExtensionAPI } from "@earendil-works/pi-coding-agent";
+import { createSdkMcpServer, type SDKUserMessage } from "@anthropic-ai/claude-agent-sdk";
 import type { Base64ImageSource, ContentBlockParam, MessageParam } from "@anthropic-ai/sdk/resources";
 import { PROVIDER_ID, messageContentToText } from "./convert.js";
-import { buildModels, fallbackModelForPrimaryModel, modelDisplayName, FABLE_MODEL_ID } from "./models.js";
-import { MCP_SERVER_NAME, MCP_TOOL_PREFIX, extractSkillsBlock } from "./skills.js";
+import { buildModels, modelDisplayName } from "./models.js";
+import { MCP_SERVER_NAME, MCP_TOOL_PREFIX } from "./skills.js";
 import { extractAllToolResults as _extractAllToolResults, type McpResult } from "./extract-tool-results.js";
 import { QueryContext, ctx, drainPendingToolCalls, popContext, stackDepth, pushContext, toolCallDrainCause } from "./query-state.js";
 import { teardownQuery } from "./query-teardown.js";
-import { loadConfig, normalizeEffortLevel, recordProjectTrust, registerExternalConfigResolver, type Config } from "./config.js";
+import { loadConfig, recordProjectTrust, registerExternalConfigResolver } from "./config.js";
 import { hasClaudeCredentials } from "./auth-presence.js";
 import { NATIVE_PROVIDER_UNSUPPORTED_MESSAGE, buildNativeProvider, supportsNativeProvider } from "./native-provider.js";
-import { extractAgentsAppend } from "./agents-md.js";
-import { buildPromptContextAppend } from "./prompt-context.js";
 import { jsonSchemaToZodShape } from "./typebox-to-zod.js";
 import { resolveGetModels } from "./pi-ai-compat.js";
-import { listAccountConnectors, resolveClaudeOAuth } from "./connector-inventory.js";
 // Re-exported from the extension entry point ON PURPOSE. Consuming apps
 // regenerate their vendored package.json with a CLOSED exports map
 // ({".": "./bundle/index.js"}), which makes Node reject BOTH a subpath import
@@ -40,15 +37,15 @@ export {
 export { connectorCachePath, connectorCacheScopeKey, readCachedConnectors, scopeKeyFor, writeCachedConnectors } from "./connector-cache.js";
 export { connectorServersSnapshot, primeConnectorServers } from "./connector-runtime.js";
 import { debug, diagDump, makeCliDebugOptions, moduleInstanceId } from "./debug.js";
-import { preflightClaudeExecutable, resolveClaudeExecutable, spawnClaudeCodeWithDiagnostics } from "./claude-executable.js";
+import { preflightClaudeExecutable, resolveClaudeExecutable } from "./claude-executable.js";
 import { appendIntegrityEntry, argKeys, extensionApi, markSessionForRebuild, reportToolResultMismatch, safeNotify, safeToolCallSummary, setExtensionApi, setPiUI, setSharedSession, sharedSession, type SessionState } from "./bridge-state.js";
-import { CLAUDE_BRIDGE_TOOL_ISOLATION, connectorQueryOptions, connectorWriteModeFor, connectorsEnabledFor, denyAllToolsHook, isChildExecutedTool, settingSourcesForQuery } from "./connectors.js";
-import { connectorCredentialEnv, connectorServersSnapshot, primeConnectorServers, readCredentialFile } from "./connector-runtime.js";
+import { connectorsEnabledFor, isChildExecutedTool } from "./connectors.js";
+import { primeConnectorServers } from "./connector-runtime.js";
 import { cancelScheduledSessionPersistence, restoreSharedSessionFromPi, schedulePersistSharedSession, syncSharedSession } from "./session-persistence.js";
 import { STREAM_IDLE_BACKOFF_HINT_MS, activeStreamIdleWatchdogs, buildStreamIdleTimeoutErrorMessage, createStreamIdleWatchdog, formatDurationShort, streamIdleTimeoutMsFromEnv } from "./stream-idle-watchdog.js";
-import { RATE_LIMIT_AUTO_RESUME_EVENT, RATE_LIMIT_TOKEN, formatAllowedRateLimitWarning, formatResetTimestamp, isUsageLimitMessage, uniqueNonEmptyLines } from "./rate-limit.js";
+import { RATE_LIMIT_TOKEN, formatResetTimestamp } from "./rate-limit.js";
 import { mapToolArgs } from "./tool-mapping.js";
-import { ensureTurnStarted, finalizeCurrentStream, finalizeToolUseTurnFromMcpInvocation, noteChildExecutedToolResults, processAssistantMessage, processStreamEvent, scheduleToolUseTurnEnd, updateTurnOutputModel } from "./assistant-stream.js";
+import { finalizeCurrentStream, finalizeToolUseTurnFromMcpInvocation, scheduleToolUseTurnEnd, updateTurnOutputModel } from "./assistant-stream.js";
 import {
 	accountSessionScope,
 	classifyClaudeFailure,
@@ -59,16 +56,20 @@ import {
 	resolveClaudeAccountRouter,
 	RetryEventBuffer,
 	safeRouterCall,
-	subscriberProfileEnv,
-	type ClaudeAccountFailureKind,
 	type ClaudeAccountRoute,
-	type ClaudeAccountRouterV1,
-	type ClaudeBridgeAccountHostV1,
 } from "./account-router.js";
+import { BRIDGE_ACCOUNT_HOST } from "./account-host.js";
+import { registerBridgeCommands } from "./bridge-commands.js";
+import { consumeQuery, emitRateLimitEvent, type ClaudeAttemptFailure } from "./consume-query.js";
+import { buildClaudeQueryOptions } from "./query-options.js";
+import { sdkQueryFactory } from "./sdk-query.js";
 
 // Re-exports: the module decomposition must not change the bundle entry's
 // public surface — unit tests and downstream consumers import these from
 // bundle/index.js.
+export { probeClaudeAccountProfile } from "./account-host.js";
+export { __testSetSdkQueryFactory } from "./sdk-query.js";
+export { resolveConfiguredEffort } from "./query-options.js";
 export { classifyClaudeExecutableBytes, preflightClaudeExecutable, resolveClaudeExecutable, spawnClaudeCodeWithDiagnostics, wrapClaudeSpawnErrorForSdk, type ClaudeExecutableFileType, type ClaudeExecutablePreflightResult } from "./claude-executable.js";
 export { __testGetBridgeIntegrityState, __testSetBridgeIntegrityState, INTEGRITY_CUSTOM_TYPE, appendIntegrityEntry, reportToolResultMismatch } from "./bridge-state.js";
 export { CONNECTOR_CALL_CUSTOM_TYPE, connectorResultByteSize, flushConnectorCallAudit, recordConnectorCallResult, setConnectorCallAuditSink, type ConnectorCallAuditData, type ConnectorCallAuditSink, type ConnectorCallOutcome } from "./connector-audit.js";
@@ -133,7 +134,6 @@ const newAssistantMessageEventStream: () => AssistantMessageEventStream =
 // native (pi >=0.81) upsert flow.
 const PRIMARY_INSTANCE_KEY = Symbol.for("claude-bridge:primaryInstance");
 const ACTIVE_STREAM_SIMPLE_KEY = Symbol.for("claude-bridge:activeStreamSimple");
-const COMMANDS_REGISTERED_KEY = Symbol.for("claude-bridge:commandsRegistered");
 // Deliberately NOT Symbol.for: rotation state rides options between the retry
 // re-entry and the original call within ONE module instance only.
 const ROTATION_STATE_KEY = Symbol("claude-bridge:rotationState");
@@ -156,153 +156,7 @@ type BridgeStreamOptions = SimpleStreamOptions & {
 // MODELS is buildModels(getModels("anthropic")) — projection kept in models.js.
 const MODELS = buildModels(getModels("anthropic"));
 
-type SdkQueryFactory = typeof query;
-let sdkQueryFactory: SdkQueryFactory = query;
 
-/** Test seam for exercising the real bridge retry/session orchestration without
- *  spending Claude usage. Production never calls this. */
-export function __testSetSdkQueryFactory(factory?: SdkQueryFactory): void {
-	sdkQueryFactory = factory ?? query;
-}
-
-function emitRateLimitEvent(payload: Record<string, unknown>): void {
-	try {
-		extensionApi?.events?.emit?.(RATE_LIMIT_AUTO_RESUME_EVENT, payload);
-	} catch {
-		// Cross-extension broker is best-effort only.
-	}
-}
-
-// The fastMode setting silently no-ops when Claude Code declines fast mode.
-// Surface the typed fast_mode_disabled_reason (SDK 0.3.219+) once per distinct
-// reason so an enabled-but-inert setting explains itself instead of looking
-// broken. Module-level dedup: the same reason repeats on every init message.
-let lastFastModeDisabledNoticeReason: string | null = null;
-
-const FAST_MODE_DISABLED_REASON_TEXT: Record<string, string> = {
-	disabled_by_env: "disabled by an environment variable",
-	extra_usage_disabled: "extra usage is disabled for this account",
-	free: "not available on the free plan",
-	model_not_allowed: "not available for this model",
-	network_error: "the eligibility check hit a network error",
-	not_first_party: "not available for this account type",
-	preference: "disabled by a Claude Code preference",
-	sdk_opt_in_required: "the SDK opt-in is missing",
-	unknown: "unavailable for an unknown reason",
-};
-
-function noteFastModeDisabledReason(message: unknown, bridgeConfig: Config): void {
-	if (bridgeConfig.provider?.fastMode !== true) return;
-	const reason = (message as { fast_mode_disabled_reason?: unknown }).fast_mode_disabled_reason;
-	// "pending" means the CLI is still deciding — not a verdict worth announcing.
-	if (typeof reason !== "string" || reason === "pending") return;
-	if (reason === lastFastModeDisabledNoticeReason) return;
-	lastFastModeDisabledNoticeReason = reason;
-	const text = FAST_MODE_DISABLED_REASON_TEXT[reason] ?? `unavailable (${reason})`;
-	safeNotify(`Pi Claude: fast mode is enabled in settings but Claude Code declined it — ${text}.`, "warning");
-}
-
-/** Local /usage probe for the reciprocal account-host service: the companion
- *  account manager asks the bridge (the SDK owner) to read a profile's identity
- *  and usage figures under that profile's credential scope. */
-// This is a published entry point (BRIDGE_ACCOUNT_HOST.probeProfile) and
-// `signal` is optional, so a stalled child with no caller signal would hang the
-// returned promise forever and leak the process. The internal deadline bounds
-// every call; generous next to the in-query probe's 1.5s race because a cold
-// child spawn is part of the budget here.
-const ACCOUNT_PROBE_DEADLINE_MS = 10_000;
-
-export async function probeClaudeAccountProfile(input: {
-	profile: ClaudeAccountRoute;
-	cwd: string;
-	signal?: AbortSignal;
-	/** Deadline override for tests; production callers use the default. */
-	deadlineMs?: number;
-}): Promise<{
-	identity?: { email?: string; organization?: string; subscriptionType?: string; authMethod?: string };
-	usage?: unknown;
-}> {
-	const config = loadConfig(input.cwd);
-	const claudeExecutable = resolveClaudeExecutable(config.provider?.pathToClaudeCodeExecutable);
-	if (claudeExecutable) preflightClaudeExecutable(claudeExecutable, input.cwd);
-	const probe = sdkQueryFactory({
-		prompt: "/usage",
-		options: {
-			cwd: input.cwd,
-			env: {
-				...subscriberProfileEnv(input.profile),
-				ENABLE_CLAUDEAI_MCP_SERVERS: "0",
-				DISABLE_AUTO_COMPACT: "1",
-			},
-			maxTurns: 1,
-			// bypassPermissions makes tool containment the only gate, so the probe
-			// gets BOTH layers: the standard bridge isolation lists AND a deny-all
-			// PreToolUse hook — a /usage probe has no business executing anything.
-			permissionMode: "bypassPermissions",
-			...CLAUDE_BRIDGE_TOOL_ISOLATION,
-			hooks: { PreToolUse: [{ hooks: [denyAllToolsHook()] }] },
-			...(claudeExecutable ? { pathToClaudeCodeExecutable: claudeExecutable } : {}),
-			spawnClaudeCodeProcess: spawnClaudeCodeWithDiagnostics,
-			...makeCliDebugOptions("account-probe"),
-		},
-	});
-	const onAbort = () => {
-		void probe.interrupt().catch(() => {});
-		try { probe.close(); } catch {}
-	};
-	if (input.signal?.aborted) onAbort();
-	else input.signal?.addEventListener("abort", onAbort, { once: true });
-	let controls: Promise<{
-		identity?: { email?: string; organization?: string; subscriptionType?: string; authMethod?: string };
-		usage?: unknown;
-	}> | undefined;
-	let deadlineTimer: ReturnType<typeof setTimeout> | undefined;
-	const deadline = new Promise<"deadline">((resolveDeadline) => {
-		deadlineTimer = setTimeout(() => {
-			debug(`account-probe: deadline expired for ${input.profile.label}; killing probe child`);
-			onAbort();
-			resolveDeadline("deadline");
-		}, input.deadlineMs ?? ACCOUNT_PROBE_DEADLINE_MS);
-		deadlineTimer.unref?.();
-	});
-	const consume = (async () => {
-		for await (const message of probe) {
-			if (message.type === "system" && (message as any).subtype === "init" && !controls) {
-				controls = Promise.allSettled([
-					probe.accountInfo(),
-					probe.usage_EXPERIMENTAL_MAY_CHANGE_DO_NOT_RELY_ON_THIS_API_YET(),
-				]).then(([identityResult, usageResult]) => ({
-					...(identityResult.status === "fulfilled" ? { identity: {
-						email: identityResult.value.email,
-						organization: identityResult.value.organization,
-						subscriptionType: identityResult.value.subscriptionType,
-					} } : {}),
-					...(usageResult.status === "fulfilled" ? { usage: usageResult.value } : {}),
-				}));
-			}
-		}
-		return controls ? await controls : {};
-	})();
-	try {
-		// The race — not close() alone — is what guarantees the promise settles: a
-		// truly wedged child can survive close(), and then the for-await above
-		// never ends.
-		const result = await Promise.race([consume, deadline]);
-		return result === "deadline" ? {} : result;
-	} finally {
-		clearTimeout(deadlineTimer);
-		input.signal?.removeEventListener("abort", onAbort);
-		probe.close();
-		// After a deadline kill the consumer may still reject; that outcome is
-		// already accounted for.
-		void consume.catch(() => {});
-	}
-}
-
-const BRIDGE_ACCOUNT_HOST: ClaudeBridgeAccountHostV1 = {
-	version: 1,
-	probeProfile: probeClaudeAccountProfile,
-};
 
 // Pi doesn't pass tool results directly — it appends them to the context and calls
 // the provider again. Thin wrapper over extract-tool-results.js that adds per-turn
@@ -542,32 +396,6 @@ function buildMcpServers(tools: Tool[], queryCtx: QueryContext): Record<string, 
 	return { [MCP_SERVER_NAME]: server };
 }
 
-// --- Effort level mapping ---
-// Pi reasoning levels → CC SDK effort levels
-
-const REASONING_TO_EFFORT: Record<string, EffortLevel> = {
-	minimal: "low", low: "low", medium: "medium", high: "high", xhigh: "max", max: "max",
-};
-
-function normalizeEffortOverrideModelKey(value: string): string {
-	const key = value.trim().toLowerCase();
-	return key.startsWith(`${PROVIDER_ID}/`) ? key.slice(PROVIDER_ID.length + 1) : key;
-}
-
-export function resolveConfiguredEffort(
-	modelId: string,
-	reasoningEffort: EffortLevel | undefined,
-	providerConfig?: Config["provider"],
-): EffortLevel | undefined {
-	const target = normalizeEffortOverrideModelKey(modelId);
-	for (const [key, rawEffort] of Object.entries(providerConfig?.modelEffortOverrides ?? {})) {
-		const normalizedKey = normalizeEffortOverrideModelKey(key);
-		if (normalizedKey !== "*" && normalizedKey !== target) continue;
-		const effort = normalizeEffortLevel(rawEffort) as EffortLevel | undefined;
-		if (effort) return effort;
-	}
-	return (normalizeEffortLevel(providerConfig?.forceEffort) as EffortLevel | undefined) ?? reasoningEffort;
-}
 
 // --- Provider: streaming function ---
 //
@@ -584,253 +412,6 @@ export function resolveConfiguredEffort(
 // currentPiStream, so any leftover messages hit the `!ctx().currentPiStream` guard
 // in consumeQuery and are skipped before resetTurnState runs.
 
-/** Background consumer: iterates the SDK generator, pushing events to currentPiStream.
- *  Runs until the query ends. Per turn, the SDK yields stream_events (deltas), then
- *  an assistant message (completed blocks). On tool_use, the stream is ended by
- *  whichever path handles it first (processStreamEvent or processAssistantMessage),
- *  and the MCP handler blocks the generator until pi delivers the tool result. */
-interface ClaudeAttemptFailure {
-	kind?: ClaudeAccountFailureKind;
-	message: string;
-	rateLimitInfo?: Record<string, unknown>;
-}
-
-interface ConsumeQueryResult {
-	capturedSessionId?: string;
-	failure?: ClaudeAttemptFailure;
-}
-
-async function consumeQuery(
-	sdkQuery: ReturnType<typeof query>,
-	// The CAPTURED context of the query being consumed, never the live ctx():
-	// an MCP tool can push a reentrant subagent context while this iterator is
-	// suspended, and reading live state then would consult the WRONG query — a
-	// recovered success could retain its failure and surface an error, or a
-	// child session id could stamp the subagent's context.
-	queryCtx: QueryContext,
-	customToolNameToPi: Map<string, string>,
-	model: Model<any>,
-	bridgeConfig: Config,
-	wasAborted: () => boolean,
-	account?: ClaudeAccountRoute,
-	router?: ClaudeAccountRouterV1,
-	// Mirror of the held failure for the caller's .catch: the SDK iterator can
-	// THROW after the failure-signal message (a rejected rate_limit_event is the
-	// known case), and the rejection loses this function's return value. Without
-	// the mirror the catch re-classifies from the thrown error, misses
-	// rateLimitInfo, and double-counts the router cooldown.
-	attemptFailureBox?: { failure?: ClaudeAttemptFailure },
-): Promise<ConsumeQueryResult> {
-	let capturedSessionId: string | undefined;
-	let failure: ClaudeAttemptFailure | undefined;
-	let accountProbe: Promise<void> | undefined;
-	const holdFailure = (next: ClaudeAttemptFailure | undefined): void => {
-		failure = next;
-		if (attemptFailureBox) attemptFailureBox.failure = next;
-	};
-
-	for await (const message of sdkQuery) {
-		if (wasAborted()) break;
-		activeStreamIdleWatchdogs.get(queryCtx)?.noteChunk();
-		if (account) {
-			debug("consumeQuery: managed message", JSON.stringify({
-				type: message.type,
-				subtype: (message as any).subtype,
-				error: (message as any).error,
-				eventType: (message as any).event?.type,
-				deltaType: (message as any).event?.delta?.type,
-				contentType: (message as any).event?.content_block?.type,
-			}));
-		}
-		if (!queryCtx.turnOutput) continue;
-		// Only RENDERING needs a live Pi stream. Failure metadata and
-		// child-executed tool results must be captured even when a tool-use turn
-		// boundary has nulled the stream — skipping them there dropped terminal
-		// failure classification and audited late connector results "unobserved".
-		const streamLive = Boolean(queryCtx.currentPiStream);
-
-		switch (message.type) {
-			case "stream_event":
-				if (!streamLive) break;
-				processStreamEvent(message, customToolNameToPi, model, queryCtx);
-				break;
-			case "assistant": {
-				// Claude Code emits a synthetic assistant text block carrying friendly
-				// rate/auth error copy before the SDK throws. On a managed attempt it
-				// is not model output: forwarding it would commit the stream and make
-				// safe pre-output account failover impossible, so hold it as failure
-				// metadata (with or without a live stream). A legacy attempt renders
-				// it exactly as before.
-				const sdkError = (message as any).error;
-				if (sdkError && account) {
-					if (!failure) holdFailure({ kind: classifyClaudeFailure(sdkError), message: String(sdkError) });
-					break;
-				}
-				if (!streamLive && !queryCtx.turnSawToolCall) break;
-				processAssistantMessage(message, model, customToolNameToPi, queryCtx);
-				break;
-			}
-			case "result":
-				// A failure signal followed by a result whose visible output already
-				// committed (e.g. the SDK's fallback-model reroute recovering after a
-				// rejected rate limit) means the query ultimately SUCCEEDED: the
-				// failure was informational and must neither surface an error after
-				// the answer nor skip session persistence.
-				if (failure && message.subtype === "success" && queryCtx.committedOutput) {
-					debug(`consumeQuery: clearing informational ${failure.kind ?? "unclassified"} failure — query recovered with committed output`);
-					holdFailure(undefined);
-				}
-				// The SDK can label the synthetic friendly error carrier as a
-				// successful result immediately before its iterator throws. Once a
-				// managed attempt holds a terminal failure signal, that text is still
-				// error metadata, not assistant output.
-				if (account && failure) break;
-				if (!queryCtx.turnSawStreamEvent && message.subtype === "success") {
-					if (!streamLive) break;
-					const text = message.result || "";
-					// The no-stream-events assistant fallback may have already rendered
-					// this exact text (it does not set turnSawStreamEvent) — re-pushing
-					// it here is the other half of the duplicated-output bug.
-					if (queryCtx.turnBlocks.some((b: any) => b.type === "text" && b.text === text)) {
-						debug("consumeQuery: result text already rendered by assistant fallback; skipping duplicate");
-						break;
-					}
-					ensureTurnStarted(queryCtx);
-					queryCtx.turnBlocks.push({ type: "text", text });
-					const idx = queryCtx.turnBlocks.length - 1;
-					queryCtx.currentPiStream?.push({ type: "text_start", contentIndex: idx, partial: queryCtx.turnOutput });
-					queryCtx.currentPiStream?.push({ type: "text_delta", contentIndex: idx, delta: text, partial: queryCtx.turnOutput });
-					queryCtx.currentPiStream?.push({ type: "text_end", contentIndex: idx, content: text, partial: queryCtx.turnOutput });
-				} else if (message.subtype !== "success") {
-					const errorLines = Array.isArray((message as any).errors) ? uniqueNonEmptyLines((message as any).errors) : [];
-					const errors = errorLines.length > 0 ? errorLines.join("\n") : String((message as any).result || message.subtype || "Claude Code request failed");
-					const usageLimit = isUsageLimitMessage(message);
-					if (!failure || !failure.rateLimitInfo) {
-						holdFailure({ kind: usageLimit ? "rate-limit" : classifyClaudeFailure(errors), message: errors });
-					}
-					// Managed attempts keep terminal error copy buffered as metadata so
-					// a pre-output failure can move to another subscription profile.
-					if (account) break;
-					if (usageLimit) {
-						// isUsageLimitMessage matches the CLI's own usage-limit copy (SDK
-						// USAGE_LIMIT_ERROR_PREFIXES). Surface it immediately, exactly as
-						// before, and suppress the SDK's raw follow-up throw.
-						queryCtx.handledTerminalError = true;
-						queryCtx.turnOutput.stopReason = "error";
-						queryCtx.turnOutput.errorMessage = errors;
-						queryCtx.currentPiStream?.push({ type: "error", reason: "error", error: queryCtx.turnOutput });
-						queryCtx.currentPiStream?.end();
-						queryCtx.currentPiStream = null;
-					}
-					// Other non-success subtypes (error_max_turns,
-					// error_during_execution) surface at completion via the held
-					// failure — an explicit error event where these turns previously
-					// ended silently. Session persistence and deferred replay still run.
-				}
-				break;
-			case "system":
-				if (!streamLive) break;
-				if ((message as any).subtype === "init" && (message as any).session_id) {
-					capturedSessionId = (message as any).session_id;
-					// Also on this message's query context, so the connector-call audit
-					// trail can name the child session that executed a call — including
-					// from the teardown flush, which runs outside this function's scope.
-					queryCtx.childSessionId = capturedSessionId;
-					noteFastModeDisabledReason(message, bridgeConfig);
-					if (account && router && !accountProbe) {
-						accountProbe = Promise.allSettled([
-							sdkQuery.accountInfo().then((info) => router.recordIdentity(account.profileId, {
-								email: info.email,
-								organization: info.organization,
-								subscriptionType: info.subscriptionType,
-							})),
-							sdkQuery.usage_EXPERIMENTAL_MAY_CHANGE_DO_NOT_RELY_ON_THIS_API_YET()
-								.then((usage) => router.recordUsage(account.profileId, usage)),
-						]).then((results) => {
-							const labels = ["recordIdentity", "recordUsage"];
-							results.forEach((result, i) => {
-								if (result.status === "rejected") debug(`consumeQuery: account probe ${labels[i]} rejected:`, result.reason);
-							});
-						});
-					}
-				} else if ((message as any).subtype === "model_refusal_fallback") {
-					const originalModel = (message as any).original_model;
-					const fallbackModel = (message as any).fallback_model;
-					updateTurnOutputModel(fallbackModel, queryCtx);
-					debug("consumeQuery: model_refusal_fallback", JSON.stringify({ originalModel, fallbackModel }));
-					// Notify only for reroutes we configured, so an unexpected pairing from
-					// Claude Code is still logged above but not announced as one of ours.
-					if (typeof fallbackModel === "string" && typeof originalModel === "string" && fallbackModelForPrimaryModel(originalModel) === fallbackModel) {
-						safeNotify(
-							`Pi Claude switched ${modelDisplayName(originalModel)} to ${modelDisplayName(fallbackModel)} after Claude Code safety fallback.`,
-							"info",
-						);
-					}
-				}
-				break;
-			case "user":
-				// Mostly the SDK echoing the prompt back — nothing to render. The one
-				// thing worth reading is a child-executed tool's real result, which
-				// arrives here and nowhere else — including AFTER a tool-use turn
-				// boundary nulled the stream (noteChildExecutedToolResults is
-				// side-effect-free on the Pi stream).
-				noteChildExecutedToolResults(message, queryCtx);
-				break;
-			case "rate_limit_event": {
-				if (!streamLive) break;
-				const info = (message as any).rate_limit_info as Record<string, unknown> | undefined;
-				debug("consumeQuery: rate_limit_event", JSON.stringify(info).slice(0, 300));
-				if (info?.status === "rejected") {
-					const rateLimitType = rateLimitTypeFromInfo(info);
-					const resetAt = rateLimitResetFromInfo(info);
-					const resetAtMs = rateLimitResetMs(info);
-					const reason = `${rateLimitType ?? "unknown"} rate limit`;
-					if (account && router) {
-						// Rotation may still recover this request, so hold the rejection
-						// as failure metadata and teach the router the reset time now.
-						// Surfacing (event + toast) happens once in surfaceFailure — only
-						// if the attempt is not replayed on another profile.
-						holdFailure({ kind: "rate-limit", message: reason, rateLimitInfo: info });
-						safeRouterCall("recordRateLimit", () => router.recordRateLimit(account.profileId, info, model.id));
-					} else {
-						// Legacy: notify once and set NO failure state. The SDK's own
-						// fallback-model path may still stream a successful recovery, and
-						// that turn must complete exactly like any other success.
-						const resetsAt = formatResetTimestamp(resetAtMs ?? resetAt);
-						emitRateLimitEvent({
-							model: model.id,
-							provider: model.provider,
-							rateLimitType,
-							reason,
-							resetAt,
-							...(Number.isFinite(resetAtMs) ? { resetAtMs } : {}),
-							source: "claude-bridge",
-							status: "rejected",
-						});
-						safeNotify(`${RATE_LIMIT_TOKEN} Claude ${reason} hit — resets ${resetsAt}`, "warning");
-					}
-				} else if (info?.status === "allowed_warning") {
-					const warning = formatAllowedRateLimitWarning(info);
-					if (warning) safeNotify(warning, "warning");
-					else debug("consumeQuery: suppressed low/ambiguous allowed_warning rate_limit_event", JSON.stringify(info).slice(0, 300));
-				}
-				break;
-			}
-			default:
-				debug("consumeQuery: unhandled SDK message type", message.type);
-				break;
-		}
-	}
-
-	if (accountProbe) {
-		await Promise.race([
-			accountProbe,
-			new Promise<void>((resolve) => setTimeout(resolve, 1_500)),
-		]);
-	}
-	debug(`consumeQuery: for-await loop exited, wasAborted=${wasAborted()}, capturedSessionId=${capturedSessionId?.slice(0, 8) ?? "none"}, failure=${failure?.kind ?? "none"}`);
-	return { capturedSessionId, failure };
-}
 
 // Claim the primary-instance token for this module instance if unclaimed, and
 // report whether this instance is the primary. First-loaded instance wins,
@@ -1239,108 +820,27 @@ export function streamClaudeAgentSdk(model: Model<any>, context: Context, option
 		? wrapPromptStream(promptBlocks)
 		: promptText;
 	const mcpServers = buildMcpServers(mcpTools, ctx());
-	// Whether to expose the Claude account's claude.ai cloud MCP connectors
-	// (Gmail/Calendar/Drive). Enabled via env or config; drives setting-sources,
-	// tool isolation, and the ENABLE_CLAUDEAI_MCP_SERVERS child-env gate below.
-	const enableCloudMcp = connectorsEnabledFor(bridgeConfig);
-	// Connector WRITE control: read-only by default (writes denied); the one-shot
-	// approved-write executor sets CLAUDE_BRIDGE_CONNECTOR_WRITE=allow / config.
-	const connectorWriteMode = connectorWriteModeFor(bridgeConfig);
-	// Declare the account's connected connectors explicitly so `alwaysLoad` can
-	// hold startup until they attach — otherwise the turn-1 manifest is built
-	// before the CLI has fetched them (vstack#832).
-	const connectorServers = enableCloudMcp ? connectorServersSnapshot(accountScope.claudeConfigDir) : {};
-	const appendSystemPrompt = providerSettings.appendSystemPrompt !== false;
-	const agentsAppend = appendSystemPrompt ? extractAgentsAppend() : undefined;
-	const skillsAppend = appendSystemPrompt ? extractSkillsBlock(context.systemPrompt) : undefined;
-	const promptContextAppend = buildPromptContextAppend(context.systemPrompt, cwd, bridgeConfig.promptContext ?? {});
-	const appendParts = [agentsAppend, skillsAppend, promptContextAppend.text].filter((part): part is string => Boolean(part));
-	const systemPromptAppend = appendParts.length > 0 ? appendParts.join("\n\n") : undefined;
-
-	// MCP auto-loading suppression: with appendSystemPrompt=true (default), the
-	// SDK uses isolation mode and avoids filesystem settings. If users turn that
-	// off, load user/project settings but pass --strict-mcp-config so Claude Code
-	// ignores auto-discovered filesystem MCP servers while Pi owns tool execution.
-	// Connectors mode needs settings resolution ON but restricted to USER scope
-	// only — project/local settings files can smuggle `env`/`apiKeyHelper` from
-	// a hostile checkout (vstack#990). Full rationale on settingSourcesForQuery.
-	const settingSources: SettingSource[] | undefined = settingSourcesForQuery(
-		enableCloudMcp, appendSystemPrompt, providerSettings.settingSources);
-	const strictMcpConfigEnabled = !appendSystemPrompt && providerSettings.strictMcpConfig !== false;
-	// Prefer the model's own thinkingLevelMap when present (pi-ai 0.72+ ships
-	// per-model overrides — e.g. opus-4-7 wants xhigh→xhigh, not xhigh→max).
-	// Fall back to our generic table for older pi-ai or unmapped levels.
-	const requestedEffort = options?.reasoning
-		? ((queryModel as any).thinkingLevelMap?.[options.reasoning] as EffortLevel | undefined)
-			?? REASONING_TO_EFFORT[options.reasoning]
-		: undefined;
-	const effort = resolveConfiguredEffort(queryModel.id, requestedEffort, providerSettings);
-
-	const extraArgs: Record<string, string | null> = {};
-	// Opus 4.7 defaults thinking.display to "omitted" (empty thinking text in stream).
-	// Force summarized so thinking_delta events arrive. See anthropics/claude-agent-sdk-python#830.
-	// Deliberately the raw flag, NOT the typed `thinking` option: every non-disabled
-	// ThinkingConfig also emits `--thinking adaptive` or `--max-thinking-tokens`
-	// (verified in sdk.mjs flag mapping), so the typed form cannot set display
-	// without overriding the model's thinking mode alongside our `--effort`.
-	if (effort) extraArgs["thinking-display"] = "summarized";
-	// With a managed Fable pool, let every account's model-scoped allowance run
-	// out (rotation) before changing models — the CLI's own Opus fallback would
-	// silently skip accounts whose Fable quota is still available. Once the
-	// router explicitly selects Opus, its normal Opus→4.8 safety fallback is
-	// back on.
-	const fallbackModel = account && model.id === FABLE_MODEL_ID && queryModel.id === model.id
-		? undefined
-		: fallbackModelForPrimaryModel(queryModel.id);
-
-	// Suppress claude.ai cloud MCP servers (Figma/Canva/etc. auto-discovered via OAuth
-	// when the user is logged into Anthropic). These are a separate code path from
-	// filesystem MCP and are NOT blocked by --strict-mcp-config or settingSources=undefined.
-	// The native CC binary gates them on env var ENABLE_CLAUDEAI_MCP_SERVERS: setting it
-	// to "0"/"false"/"no"/"off" makes the loader return early before any cloud fetch.
-	// DISABLE_AUTO_COMPACT=1: pi owns context-management and propagates its own
-	// /compact via session_compact (see handler in default export). Letting CC
-	// also autocompact would double-flush the prompt cache and races pi's
-	// threshold with CC's, including CC's anti-thrashing guard (issue #8).
-	// Manual /compact in CC still works (we never invoke it).
-	// When connectors are enabled, allow claude.ai cloud MCP servers so the
-	// authenticated account's Gmail/Calendar/Drive tools load. Default stays "0".
-	const childEnv = {
-		...(account ? subscriberProfileEnv(account) : process.env),
-		ENABLE_CLAUDEAI_MCP_SERVERS: enableCloudMcp ? "1" : "0",
-		DISABLE_AUTO_COMPACT: "1",
-	};
-	const queryOptions: NonNullable<Parameters<typeof query>[0]["options"]> = {
+	// Pure SDK query-option assembly — see buildClaudeQueryOptions for the
+	// connector, prompt-append, setting-source, effort, and env rationale.
+	const built = buildClaudeQueryOptions({
 		cwd,
-		model: queryModel.id,
-		env: childEnv,
-		...connectorQueryOptions(enableCloudMcp, connectorWriteMode),
-		permissionMode: "bypassPermissions",
-		includePartialMessages: true,
-		...(fallbackModel ? { fallbackModel } : {}),
-		...(providerSettings.fastMode ? { settings: { fastMode: true } } : {}),
-		systemPrompt: {
-			type: "preset", preset: "claude_code",
-			append: systemPromptAppend ? systemPromptAppend : undefined,
-		},
-		extraArgs,
-		...(strictMcpConfigEnabled ? { strictMcpConfig: true } : {}),
-		...(effort ? { effort } : {}),
-		...(settingSources ? { settingSources } : {}),
-		...(mcpServers || Object.keys(connectorServers).length > 0
-			? { mcpServers: { ...(mcpServers ?? {}), ...connectorServers } as NonNullable<Parameters<typeof query>[0]["options"]>["mcpServers"] }
-			: {}),
-		...(resumeSessionId ? { resume: resumeSessionId } : {}),
-		...(claudeExecutable ? { pathToClaudeCodeExecutable: claudeExecutable } : {}),
-		spawnClaudeCodeProcess: spawnClaudeCodeWithDiagnostics,
-		...makeCliDebugOptions("provider"),
-	};
+		requestedModel: model,
+		queryModel,
+		account,
+		bridgeConfig,
+		systemPrompt: context.systemPrompt,
+		reasoning: options?.reasoning,
+		resumeSessionId,
+		mcpServers,
+		claudeExecutable,
+	});
+	const { queryOptions } = built;
 
 	debug("provider: fresh query",
 		`model=${queryModel.id} requested=${model.id} msgs=${context.messages.length} tools=${mcpTools.length}`,
-		`resume=${resumeSessionId?.slice(0, 8) ?? "none"} effort=${effort ?? "default"} account=${account?.label ?? "legacy"}`,
-		`fallback=${fallbackModel ?? "none"}`,
-		`appendSys=${appendSystemPrompt} promptCtx=${promptContextAppend.labels.join(",") || "none"} strictMcp=${strictMcpConfigEnabled} fastMode=${providerSettings.fastMode === true} connectors=${enableCloudMcp}`,
+		`resume=${resumeSessionId?.slice(0, 8) ?? "none"} effort=${built.effort ?? "default"} account=${account?.label ?? "legacy"}`,
+		`fallback=${built.fallbackModel ?? "none"}`,
+		`appendSys=${built.appendSystemPrompt} promptCtx=${built.promptContextLabels.join(",") || "none"} strictMcp=${built.strictMcpConfigEnabled} fastMode=${providerSettings.fastMode === true} connectors=${built.enableCloudMcp}`,
 		`claudeExec=${claudeExecutablePreflight ? `${claudeExecutablePreflight.fileType}:${claudeExecutablePreflight.path}` : "sdk-default"}`,
 		`prompt=${promptText.slice(0, 60)}${promptBlocks ? " [+images]" : ""}`);
 
@@ -1750,81 +1250,6 @@ export function streamClaudeAgentSdk(model: Model<any>, context: Context, option
 		});
 
 	return stream;
-}
-
-function commandCwd(ctx: unknown): string {
-	const value = (ctx as { cwd?: unknown })?.cwd;
-	return typeof value === "string" && value.length > 0 ? value : process.cwd();
-}
-
-async function tryOpenExtensionManagerSettings(ctx: { ui: ExtensionUIContext }): Promise<boolean> {
-	const host = globalThis as unknown as Record<PropertyKey, unknown>;
-	const openQuickSettings = host[Symbol.for("vstack.pi.extension-manager.open-quick-settings")];
-	if (typeof openQuickSettings !== "function") return false;
-	try {
-		await (openQuickSettings as (ctx: unknown, hint?: string) => Promise<void>)(ctx, "@vanillagreen/pi-claude-bridge");
-		return true;
-	} catch {
-		return false;
-	}
-}
-
-function showBridgeStatus(ctx: { ui: ExtensionUIContext; cwd?: string }): void {
-	const config = loadConfig(commandCwd(ctx));
-	ctx.ui.notify([
-		`Pi Claude: ${config.enabled === false ? "disabled" : "enabled"}`,
-		"Claude account billing settings (including Extra Usage) are managed in Claude.",
-	].join("\n"), "info");
-}
-
-// Deterministic connector enumeration for the host app (vstack#838). Reports the
-// failure reason rather than an empty list, so "no connectors" and "could not
-// check" stay distinguishable.
-async function reportConnectorInventory(ctx: {
-	ui: ExtensionUIContext;
-	model?: Model<any>;
-	sessionManager?: { getSessionId?: () => string };
-}): Promise<void> {
-	// With a router active, enumerate the CURRENT route's account rather than
-	// whatever the process env points at.
-	const account = ctx.model
-		? resolveClaudeAccountRouter()?.current(ctx.model.id, ctx.sessionManager?.getSessionId?.())
-		: undefined;
-	const credentials = resolveClaudeOAuth(readCredentialFile, connectorCredentialEnv(account ? accountSessionScope(account).claudeConfigDir : undefined));
-	if (!credentials) {
-		ctx.ui.notify("Pi Claude: no Claude OAuth credentials found — cannot enumerate connectors.", "error");
-		return;
-	}
-	const inventory = await listAccountConnectors({ credentials });
-	if (!inventory.ok) {
-		ctx.ui.notify(`Pi Claude: connector enumeration failed — ${inventory.reason}`, "error");
-		return;
-	}
-	if (inventory.connectors.length === 0) {
-		ctx.ui.notify("Pi Claude: this account has no connectors installed.", "info");
-		return;
-	}
-	const names = inventory.connectors.map((c) => c.name).join(", ");
-	ctx.ui.notify(`Pi Claude: ${inventory.connectors.length} connector(s) installed — ${names}`, "info");
-}
-
-function registerBridgeCommands(pi: ExtensionAPI): void {
-	const guard = pi as unknown as Record<PropertyKey, unknown>;
-	if (guard[COMMANDS_REGISTERED_KEY]) return;
-	guard[COMMANDS_REGISTERED_KEY] = true;
-
-	pi.registerCommand("pi-claude", {
-		description: "Open Pi Claude settings/status",
-		handler: async (args: string, ctx) => {
-			if (args.trim()) ctx.ui.notify("Unknown /pi-claude argument.", "warning");
-			if (await tryOpenExtensionManagerSettings(ctx)) return;
-			showBridgeStatus(ctx);
-		},
-	});
-	pi.registerCommand("pi-claude:connectors", {
-		description: "List the Claude account's installed claude.ai connectors",
-		handler: async (_args: string, ctx) => reportConnectorInventory(ctx),
-	});
 }
 
 // --- Extension registration ---
